Norm: parse - parse部分を分割
